Adicionando verificações para saber se o usuário tem permissão de realizar determinadas ações.

+++ app/Http/Controllers/Admin/GruposController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\admin;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Repositories\GrupoRepository;
+use Illuminate\Support\Facades\Gate;
 
 class GruposController extends Controller
 {
@@ -14,20 +15,35 @@ class GruposController extends Controller
     public function __construct(GrupoRepository $grupoRepository) {
         $this->grupoRepository = $grupoRepository;
     }
-
+    
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index(){
+        if(Gate::denies('grupo'))
+            return redirect()->back()->with('error', 'Ops! Acesso negado.');
+
+        $columns = json_encode($this->getColunas());
         $breadcrumb = json_encode([
             ["titulo"=>"Home", "url" =>route('home')],
             ["titulo"=>"Grupos", "url" =>""]
-        ]);
-        return view('master.grupo.indexGrupo', compact('breadcrumb'));
+            ]);
+        return view('master.grupo.indexGrupo', compact('breadcrumb', 'columns'));           
     }
 
+    private function getColunas() {
+        $columns = array(["field"=>"idGrupo", "hidden" =>true]);
+        if (Gate::allows('excluir_Grupo'))
+            array_push($columns,["field"=>"deletar", "label" =>'', "width"=> '50px', "sortable"=>false]);
+
+        if (Gate::allows('editar_Grupo'))
+            array_push($columns,["field"=>"editar", "label" =>'', "width"=> '50px', "sortable"=>false]);
+            
+        array_push($columns,["field"=>"nomeGrupo", "label" =>"Grupo abc"]);
+        return $columns;
+    }
     /**
      * Retorna os dados paginados e com filtro
      * @param String $campo
@@ -44,6 +60,9 @@ class GruposController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create(){
+        if(Gate::denies('incluir_Grupo'))
+            return redirect()->back()->with('error', 'Ops! Acesso negado.');
+
         $breadcrumb = json_encode([
             ["titulo"=>"Home", "url" =>route('home')],
             ["titulo"=>"Grupos", "url" =>route('grupos.index')],
@@ -60,6 +79,8 @@ class GruposController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('incluir_Grupo')) 
+            return redirect()->back()->with('error', 'Ops! Acesso negado.');
         $this->valida($request);
         $this->grupoRepository->createGrupo($request->all());
         return redirect()->route('grupos.index')->with('success', 'Grupo criado com sucesso!');
@@ -87,6 +108,9 @@ class GruposController extends Controller
      */
     public function edit($id)
     {
+        if (Gate::denies('editar_Grupo')) 
+            return redirect()->back()->with('error', 'Ops! Acesso negado.');
+        
         $grupo = $this->grupoRepository->find($id, 'funcoes');
         $breadcrumb = json_encode([
             ["titulo"=>"Home", "url" =>route('home')],
@@ -105,7 +129,8 @@ class GruposController extends Controller
      */
     public function update(Request $request, $id)
     {
-        
+        if (Gate::denies('editar_Grupo')) 
+            return redirect()->back()->with('error', 'Ops! Acesso negado.');
         $this->valida($request,$id);
         $this->grupoRepository->updateGrupo($request->all(), $id);
        // Redireciona para a página da listagem dos grupos juntamente com a mensagem de sucesso.
@@ -120,6 +145,9 @@ class GruposController extends Controller
      */
     public function destroy($id)
     {
+        if (Gate::denies('excluir_Grupo')) 
+            return redirect()->back()->with('error', 'Ops! Acesso negado.');
+        
         if($this->grupoRepository->delete($id)) {
             return redirect()->route('grupos.index')->with('success', 'Grupo excluído  com sucesso!');
         }
